fix(list): incorrect item spacing with leading/trailing variants in RTL
Currently if developers use the `mdc-list-item` in RTL mode with
`--with-trailing-meta`, `--with-trailing-icon` et all, the horizontal
spacing on the start breaks due to the `reflexive-property` incorrectly
using `auto` instead of the default `16px` for the start side. The `auto`
value accidentally overrides the already-set `16px` horizontal padding
resulting in no side padding at the start.